+++ src/main/java/com/atlauncher/data/modpacksch/ModpacksChPackVersionModsManifest.java
@@ -17,12 +17,22 @@
  */
 package com.atlauncher.data.modpacksch;
 
-import com.google.gson.annotations.SerializedName;
+import java.util.List;
 
-public enum ModpacksChPackVersionManifectTargetType {
-    @SerializedName("modloader")
-    MODLOADER,
-
-    @SerializedName("game")
-    GAME
+public class ModpacksChPackVersionModsManifest {
+    public ModpacksChPackVersionSpecs specs;
+    public List<ModpacksChPackVersionManifestTarget> targets;
+    public int installs;
+    public int plays;
+    public int refreshed;
+    public String changelog;
+    public int parent;
+    public String notification;
+    public List<Object> links;
+    public String status;
+    public int id;
+    public String name;
+    public String type;
+    public int updated;
+    public List<ModpacksChPackVersionManifestMod> mods;
 }